+++ src/cli.js
@@ -20,9 +20,9 @@ import {
   skillsValidator,
   skillsSideChain,
 } from './saihubot-cli-skill-eth';
-import {skillEth2Stats, skillBeaconLastBlock} from './saihubot-cli-skill-eth2';
+import {skillEth2Stats, skillBeaconLastBlock, skillGetValidatorBlance} from './saihubot-cli-skill-eth2';
 import {skillsGas} from './saihubot-cli-skill-ethgas';
-import {skillGetBlance, skillGetValidatorBlance} from './saihubot-cli-skill-balance';
+import {skillGetBlance} from './saihubot-cli-skill-balance';
 import {skills as skillsXDAI} from './saihubot-cli-skill-xdai';
 import {skillChainId} from './saihubot-cli-skill-chainid';
 import {skillAwesomeEth} from './saihubot-skill-awesome';