+++ src/saihubot-cli-skill-balance.js
@@ -7,9 +7,10 @@ import Table from 'ink-table';
 import { t } from 'saihubot-cli-adapter/dist/i18n';
 
 import {getConfig, parseArg, getNodeURL, toArray} from './utils';
-import {XdaiBalances} from './saihubot-cli-skill-xdai';
 import {i18nValidator} from './i18n';
 import {tokenMap} from './token';
+import {ValidatorBalances} from './saihubot-cli-skill-eth2';
+import {XdaiBalances} from './saihubot-cli-skill-xdai';
 
 const balanceI18n = {
   'en': {
@@ -78,47 +79,6 @@ const EthBalances = ({addresses, fetch}) => {
   : (<Text>{t('query', {i18n: balanceI18n})}</Text>);
 }
 
-// support multiple validators balance by comma (without space)
-const ValidatorBalances = ({validator, fetch}) => {
-  const [balance, setBalance] = useState([]);
-  if (!validator) return null;
-
-  let validators = toArray(parseArg(validator));
-  const isOverflow = validators.length > 100;
-  const data = [];
-  useEffect(() => {
-    if (isOverflow) {
-      validators = validators.slice(0, 100);
-    }
-
-    async function fetchValidatorBalance() {
-      fetch(`https://beaconcha.in/api/v1/validator/${validators.join(',')}`)
-        .then(response => response.json())
-        .then(json => {
-          const arrData = toArray(json.data);
-          json.data && Object.values(arrData).map(validator => {
-            validator.balance && data.push({
-              Symbol: 'ETH',
-              Balance: `${Number(validator.balance)/10**9}`,
-              Index: validator.validatorindex,
-            });
-          });
-          setBalance(data);
-        });
-      }
-    validator && fetchValidatorBalance();
-  }, [validator, fetch]);
-
-  return balance.length > 0
-    ? (<>
-      <Text>{t('validatorBalance', {i18n: i18nValidator})}</Text>
-      <Text>{isOverflow ? '(Only shows the first 100 validators)' : ''}</Text>
-      <Table data={balance} />
-      <Text> </Text>
-    </>)
-    : (<Text>{validators} {t('query', {i18n: i18nValidator})}</Text>);
-}
-
 // support multiple account balance by comma (without space)
 // also shows validator balances
 const Balances = ({addresses, fetch}) => {
@@ -182,35 +142,5 @@ export const skillGetBlance = {
   },
 }
 
-/**
- * Get Validator's balance of [key].
- *
- * can pass the validator key, or pre-define the
- * SAIHUBOT_VALIDATOR environment variable
- */
-export const skillGetValidatorBlance = {
-  name: 'balance-validator',
-  help: '💰balance-(validator|eth2) - Show Validator\'s balance of [key]',
-  requirements: {
-    addons: ['fetch'],
-  },
-  rule: /(^balance-(validator|eth2) )(.*)|^balance-(validator|eth2)$/i,
-  action: function(robot, msg) {
-    let validator = '';
-    if (msg[3] === undefined) {
-      validator = getConfig('VALIDATOR', '');
-      if (!validator) {
-        robot.send(t('needAddr', {i18n: i18nValidator}));
-        robot.render();
-        return;
-      }
-    }
-
-    const data = validator || parseArg(msg[3]);
-    robot.sendComponent(<ValidatorBalances validator={data} fetch={robot.addons.fetch} />);
-    robot.render();
-  }
-}
-
-const skills = [skillGetBlance, skillGetValidatorBlance];
+const skills = [skillGetBlance];
 export {skills};