+++ src/saihubot-cli-skill-eth2.js
@@ -2,13 +2,15 @@
 
 import React, {useEffect, useState} from 'react';
 import { Text } from 'ink';
+import Table from 'ink-table';
 import AsciiBar from 'ascii-bar';
 import humanizeDuration from 'humanize-duration';
 import commaNumber from 'comma-number';
 import { t } from 'saihubot-cli-adapter/dist/i18n';
 
 import {rpcEthBalance} from './ethRpc';
-import {ethFetch} from './utils';
+import {parseArg, ethFetch, toArray} from './utils';
+import {i18nValidator} from './i18n';
 
 const ADDR = {
   ETH2_DEPOSIT: '0x00000000219ab540356cbb839cbe05303d7705fa',
@@ -195,6 +197,77 @@ export const skillBeaconLastBlock = {
   },
 }
 
-export const skillsETH2 = [skillEth2Stats, skillBeaconLastBlock];
+// support multiple validators balance by comma (without space)
+export const ValidatorBalances = ({validator, fetch}) => {
+  const [balance, setBalance] = useState([]);
+  if (!validator) return null;
+
+  let validators = toArray(parseArg(validator));
+  const isOverflow = validators.length > 100;
+  const data = [];
+  useEffect(() => {
+    if (isOverflow) {
+      validators = validators.slice(0, 100);
+    }
+
+    async function fetchValidatorBalance() {
+      fetch(`https://beaconcha.in/api/v1/validator/${validators.join(',')}`)
+        .then(response => response.json())
+        .then(json => {
+          const arrData = toArray(json.data);
+          json.data && Object.values(arrData).map(validator => {
+            validator.balance && data.push({
+              Symbol: 'ETH',
+              Balance: `${Number(validator.balance)/10**9}`,
+              Index: validator.validatorindex,
+            });
+          });
+          setBalance(data);
+        });
+      }
+    validator && fetchValidatorBalance();
+  }, [validator, fetch]);
+
+  return balance.length > 0
+    ? (<>
+      <Text>{t('validatorBalance', {i18n: i18nValidator})}</Text>
+      <Text>{isOverflow ? '(Only shows the first 100 validators)' : ''}</Text>
+      <Table data={balance} />
+      <Text> </Text>
+    </>)
+    : (<Text>{validators} {t('query', {i18n: i18nValidator})}</Text>);
+}
+
+/**
+ * Get Validator's balance of [key].
+ *
+ * can pass the validator key, or pre-define the
+ * SAIHUBOT_VALIDATOR environment variable
+ */
+export const skillGetValidatorBlance = {
+  name: 'balance-validator',
+  help: '💰balance-(validator|eth2) - Show Validator\'s balance of [key]',
+  requirements: {
+    addons: ['fetch'],
+  },
+  rule: /(^balance-(validator|eth2) )(.*)|^balance-(validator|eth2)$/i,
+  action: function(robot, msg) {
+    let validator = '';
+    if (msg[3] === undefined) {
+      validator = getConfig('VALIDATOR', '');
+      if (!validator) {
+        robot.send(t('needAddr', {i18n: i18nValidator}));
+        robot.render();
+        return;
+      }
+    }
+
+    const data = validator || parseArg(msg[3]);
+    robot.sendComponent(<ValidatorBalances validator={data} fetch={robot.addons.fetch} />);
+    robot.render();
+  }
+}
+
+export const skillsETH2 = [skillEth2Stats, skillBeaconLastBlock, skillGetValidatorBlance];
 const skills = [...skillsETH2];
 export {skills};